initial convert Blog to vertical
