+++ api/src/controllers.ts
@@ -1,9 +1,8 @@
 // codegen:start {preset: barrel, include: ./*.controllers.ts, import: default}
 import accountsControllers from "./Accounts.controllers.js"
-import blogControllers from "./Blog.controllers.js"
 import helloWorldControllers from "./HelloWorld.controllers.js"
 import operationsControllers from "./Operations.controllers.js"
 import usersControllers from "./Users.controllers.js"
 
-export { accountsControllers, blogControllers, helloWorldControllers, operationsControllers, usersControllers }
+export { accountsControllers, helloWorldControllers, operationsControllers, usersControllers }
 // codegen:end
